requirements.txt
aiogram
requests
python-dotenv
flask
